[Bronze V] Title: 이상한 기호, Time: 0 ms, Memory: 2020 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.cc
@@ -3,10 +3,10 @@ using namespace std;
 
 int main() {
 
-	int a, b;
+	long long int a, b;
 	cin >> a >> b;
 
-	int at;
+	long long int at;
 	at = (a + b) * (a - b);
 
 	cout << at;
